# docs/admin/admin_swagger.yaml
basePath: /admin/api/v1/
definitions:
  http_schema.HttpConfigRequest:
    properties:
      enable:
        type: boolean
      host:
        type: string
      port:
        type: integer
    type: object
  remote_schema.RemoteConnectConfigRequest:
    properties:
      api_server_url:
        type: string
      client_id:
        type: string
      enable:
        type: boolean
      msg_server_urls:
        items:
          type: string
        type: array
      time_stamp_check:
        type: boolean
    type: object
  schema.DiscoverSchema:
    properties:
      enabled:
        type: boolean
    type: object
  schema.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  schema.PcUserInfo:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  schema.ResponseData:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      request_id:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:2093
info:
  contact:
    email: me@czqu.net
    name: API Support
    url: https://rfu.czqu.net
  description: Remote unlock module API documentation
  termsOfService: https://rfu.czqu.net
  title: Remote Unlock Module Admin API documentation
  version: "1.0"
paths:
  /control-pc/{action}/:
    post:
      consumes:
      - application/json
      description: Control the operation of the computer according to the transmitted
        parameters
      parameters:
      - description: The type of operation（shutdown、standby、lock）
        enum:
        - shutdown
        - standby
        - lock
        in: path
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid action type
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: The operation failed
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Control  computer
  /control-pc/interface:
    get:
      consumes:
      - application/json
      description: Based on the specified IP version type, a list of valid network
        interfaces is returned
      parameters:
      - default: "4"
        description: IP Version Type (4 or 6)
        enum:
        - "4"
        - "6"
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A list of valid network interfaces is successfully returned
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: The request parameter is incorrect
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Server internal error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Returns a valid network interface
      tags:
      - Network interfaces
  /discovery/config:
    get:
      consumes:
      - application/json
      description: Get the Discovery Service configuration
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Server internal error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Get Discover Service Config
      tags:
      - Discover
    patch:
      consumes:
      - application/json
      description: Update the configuration of the Discover service with the provided
        settings.
      parameters:
      - description: New configuration settings
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/schema.DiscoverSchema'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated configuration.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid request parameters.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Update Discover Service Configuration
      tags:
      - Discover
  /discovery/restart:
    post:
      consumes:
      - application/json
      description: Restart the discover service.
      produces:
      - application/json
      responses:
        "200":
          description: Service restarted successfully.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Restart Discover Service
      tags:
      - Discover
  /http/config:
    get:
      consumes:
      - application/json
      description: Retrieve the current HTTP configuration based on the provided type.
      parameters:
      - description: Configuration type (HTTP_SERVICE_API or HTTPS_SERVICE_API)
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved configuration.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid request parameters.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Get HTTP Configuration
      tags:
      - HTTP
    patch:
      consumes:
      - application/json
      description: Partially update the HTTP configuration based on the provided type
        and settings.
      parameters:
      - description: Configuration type (HTTP_SERVICE_API or HTTPS_SERVICE_API)
        in: query
        name: type
        required: true
        type: string
      - description: Partial configuration settings
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/http_schema.HttpConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated configuration.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid request parameters.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Patch HTTP Configuration
      tags:
      - HTTP
    put:
      consumes:
      - application/json
      description: Update the HTTP configuration based on the provided type and settings.
      parameters:
      - description: Configuration type (HTTP_SERVICE_API or HTTPS_SERVICE_API)
        in: query
        name: type
        required: true
        type: string
      - description: Configuration settings
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/http_schema.HttpConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated configuration.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid request parameters.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Update HTTP Configuration
      tags:
      - HTTP
  /http/restart:
    post:
      consumes:
      - application/json
      description: Restart the server based on the provided type.
      parameters:
      - description: Service type (HTTP_SERVICE_API or HTTPS_SERVICE_API)
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Service restarted successfully.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Restart Service
      tags:
      - HTTP
  /info:
    get:
      consumes:
      - application/json
      description: Get the software version information
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Server internal error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Get Software Info
      tags:
      - System
  /interface/{ip}:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: The request is incorrect
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal errors
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Obtain the MAC address based on the IP address
  /interface/{ip}/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: success
        "400":
          description: The request is incorrect
        "500":
          description: Internal errors
      security:
      - ApiKeyAuth: []
      summary: Obtain the interface information based on the IP address
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password to obtain a JWT token.
      parameters:
      - description: User login credentials
        in: body
        name: loginData
        required: true
        schema:
          $ref: '#/definitions/schema.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid request parameters.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: User Login
      tags:
      - User
  /ping:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: success
      summary: Ping
  /power-saving:
    post:
      consumes:
      - application/json
      description: Enable or disable power saving mode. This function first records
        the setting in the database. If the database write fails, it returns immediately.
        If the database write succeeds, it sets the power saving mode. A failure in
        setting the mode does not affect the database value.
      parameters:
      - default: "true"
        description: Enable power saving mode (true or false)
        in: query
        name: enable
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Power saving mode set successfully.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid request parameters.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Set Power Saving Mode
      tags:
      - System
  /remote/config:
    get:
      consumes:
      - application/json
      description: Retrieve the current configuration for remote connections.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved configuration.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid request parameters.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Get Remote Connect Configuration
      tags:
      - Remote
    patch:
      consumes:
      - application/json
      description: Partially update the configuration for remote connections with
        the provided settings.
      parameters:
      - description: Partial configuration settings
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/remote_schema.RemoteConnectConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated configuration.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid request parameters.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Patch Remote Connect Configuration
      tags:
      - Remote
    put:
      consumes:
      - application/json
      description: Update the configuration for remote connections with the provided
        settings.
      parameters:
      - description: New configuration settings
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/remote_schema.RemoteConnectConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated configuration.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: Invalid request parameters.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Update Remote Connect Configuration
      tags:
      - Remote
  /remote/restart:
    post:
      consumes:
      - application/json
      description: Restart the remote service.
      produces:
      - application/json
      responses:
        "200":
          description: Service restarted successfully.
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schema.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Restart Remote Service
      tags:
      - Remote
  /unlock:
    post:
      parameters:
      - description: User Information
        in: body
        name: UserInfo
        required: true
        schema:
          $ref: '#/definitions/schema.PcUserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "400":
          description: The request is incorrect
          schema:
            $ref: '#/definitions/schema.ResponseData'
        "500":
          description: Internal errors
          schema:
            $ref: '#/definitions/schema.ResponseData'
      summary: Unlock your computer
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"